Added initial support for READ and DATA statements.
Lexer.cs:
Added support for reading Data token. Also we need to switch "modes" when reading a DATA statement. A DATA statement can have non-quoted strings. So I added a variable "readingData" which indicates if we are reading a DATA statement or not. Also needed to modify NextString to be able to read quoted and non-quoted strings.

mbasic.csproj:
Added Read.cs and Data.cs and data.mbas (to samples).

Parser.cs:
Added support for parsing DATA and READ statements. Also the constructor for Parser now takes a List<object>. This list is used to collect all the items found in DATA statements.

Program.cs:
Reserved the words READ and DATA. Also created a List<object> instance to pass into Parser constructor. Add a region to emit the code to generate the static list of DATA used by the READ statements.

Data.cs:
Added. A shell of a statement. It never will show up in an AST but is used to
make the parser be able to deal with DATA statements. It is a singleton class.

Read.cs:
Added. Used to represent a READ statement in an AST.

Token.cs
Add Read and Data.

BuiltIns.cs:
Fixed a problem in PrintString that resulted in zero length strings leaving a print serperator in effect.

Added AddData method for adding all of the static DATA to a list. Added ReadDouble and ReadString methods for reading from the list. Added Restore method to support RESTORE statement.

samples: Added data.mbas for testing the scenarios in DATA section of guide.
Also added testData.py for checking the results of data.mbas.