Added support for RESTORE statement. Also modified Lexer to be able to parse strings with double quotes in them.
Lexer.cs: 
Added support for double quotes in a string.

mbasic.csproj:
Added Restore.cs and read.mbas

Parser.cs:
Added support for parsing RESTORE. Needed to modify the structure used to collect the DATA elements to allow for RESTORING'ing back to any label. Fixed MorePrintList to make sure a print seperator is used between print items. 

Program.cs:
Modified the structure used for collection DATA elements to support RESTORE'ing to a label number. Reserved the word RESTORE.

For.cs:
Added a LineId to the constructor so that a label could be associated with the FOR statement (which is needed if we want to GOTO a for statement).

Statement.cs:
Added lineLabelAssigned field to make sure that MarkLabel doesn't throw an exception if lineLabel has not been assigned.

Token.cs:
Added Restore

print.mbas:
Added tests for double quotes in a string.

testPrint.py:
Added check for test added in print.mbas.

BuiltIns.cs:
Minor fix to PrintNumber to make sure a leading space is printed only after any NewLine is printed. And fixed the calculation. 

Changed the structure used to hold DATA elements. As a result I needed one common Read method to be shared by ReadDouble and ReadStrinng which would keep the pointers to the current location up to date in one place. (Probably should be its own class.)

Restore.cs
Added. Represents a Restore statement in an AST.

read.mbas:
Added tests from II-61 and II-62.

testRead.py:
Checks the results of read.mbas.