Add ArrayTest
During the creation of this test I noticed that each test has the
same Invoke method except for the name of the program. Therefore,
I pulled the guts of the Invoke method up to OutputExpectationTest
class, pass the name of the program to execute in constructor,
and just leave a Test method in each called Invoke which calls
the parent class implementation.